+++ lib/controllers/services.ts
@@ -32,7 +32,12 @@ import {
   ResponseErrorNotFound,
   ResponseSuccessJson
 } from "italia-ts-commons/lib/responses";
-import { IResponseErrorQuery, ResponseErrorQuery } from "../utils/response";
+import {
+  IResponseErrorQuery,
+  IResponseSuccessJsonIterator,
+  ResponseErrorQuery,
+  ResponseSuccessJsonIterator
+} from "../utils/response";
 
 import { NonEmptyString } from "italia-ts-commons/lib/strings";
 
@@ -41,6 +46,13 @@ import {
   IAzureUserAttributes
 } from "../utils/middlewares/azure_user_attributes";
 
+import { pick } from "italia-ts-commons/lib/types";
+import { FiscalCode } from "../api/definitions/FiscalCode";
+import { ServiceId } from "../api/definitions/ServiceId";
+import { SenderServiceModel } from "../models/sender_service";
+import { mapResultIterator } from "../utils/documentdb";
+
+import { RequiredQueryParamMiddleware } from "../utils/middlewares/required_query_param";
 import {
   checkSourceIpForHandler,
   clientIPAndCidrTuple as ipTuple
@@ -55,8 +67,21 @@ type IGetServiceHandler = (
   auth: IAzureApiAuthorization,
   clientIp: ClientIp,
   userAttributes: IAzureUserAttributes,
-  serviceId: NonEmptyString
+  serviceId: ServiceId
 ) => Promise<IGetServiceHandlerRet>;
+
+type IGetSenderServicesHandlerRet =
+  | IResponseSuccessJsonIterator<{
+      readonly serviceId: ServiceId;
+    }>
+  | IResponseErrorQuery;
+
+type IGetSenderServicesHandler = (
+  auth: IAzureApiAuthorization,
+  clientIp: ClientIp,
+  userAttributes: IAzureUserAttributes,
+  fiscalCode: FiscalCode
+) => Promise<IGetSenderServicesHandlerRet>;
 
 /**
  * Converts a retrieved service to a service that can be shared via API
@@ -116,6 +141,49 @@ export function GetService(serviceModel: ServiceModel): express.RequestHandler {
     ClientIpMiddleware,
     azureUserAttributesMiddleware,
     requiredServiceIdMiddleware
+  );
+  return wrapRequestHandler(
+    middlewaresWrap(
+      checkSourceIpForHandler(handler, (_, c, u, __) => ipTuple(c, u))
+    )
+  );
+}
+
+/**
+ * Returns the serviceId for all the Services that have sent
+ * at least one notification to the recipient with the provided fiscalCode.
+ */
+export function GetSenderServicesHandler(
+  senderServiceModel: SenderServiceModel
+): IGetSenderServicesHandler {
+  return async (_, __, ___, fiscalCode) => {
+    const retrievedServicesIterator = senderServiceModel.findSenderServicesForRecipient(
+      fiscalCode
+    );
+    const senderServicesIterator = mapResultIterator(
+      retrievedServicesIterator,
+      service => pick(["serviceId"], service)
+    );
+    return ResponseSuccessJsonIterator(senderServicesIterator);
+  };
+}
+
+/**
+ * Wraps a GetSenderServices handler inside an Express request handler.
+ */
+export function GetSenderServices(
+  serviceModel: ServiceModel,
+  senderServiceModel: SenderServiceModel
+): express.RequestHandler {
+  const handler = GetSenderServicesHandler(senderServiceModel);
+  const azureUserAttributesMiddleware = AzureUserAttributesMiddleware(
+    serviceModel
+  );
+  const middlewaresWrap = withRequestMiddlewares(
+    AzureApiAuthMiddleware(new Set([UserGroup.ApiFullProfileRead])),
+    ClientIpMiddleware,
+    azureUserAttributesMiddleware,
+    RequiredQueryParamMiddleware("recipient", FiscalCode)
   );
   return wrapRequestHandler(
     middlewaresWrap(